[OSIS-3527] move search views to separate files
	new file:   base/templates/learning_unit/search/service_course.html
	new file:   base/templates/learning_unit/search/simple.html
	modified:   base/urls.py
	deleted:    base/views/learning_units/search.py
	new file:   base/views/learning_units/search/__init__.py
	new file:   base/views/learning_units/search/borrowed.py
	new file:   base/views/learning_units/search/common.py
	new file:   base/views/learning_units/search/external.py
	new file:   base/views/learning_units/search/proposal.py
	new file:   base/views/learning_units/search/service_course.py
	new file:   base/views/learning_units/search/simple.py
