+++ base/urls.py
@@ -35,7 +35,11 @@ import base.views.learning_units.delete
 import base.views.learning_units.educational_information
 import base.views.learning_units.proposal.consolidate
 import base.views.learning_units.proposal.delete
-import base.views.learning_units.search
+import base.views.learning_units.search.borrowed
+import base.views.learning_units.search.external
+import base.views.learning_units.search.proposal
+import base.views.learning_units.search.service_course
+import base.views.learning_units.search.simple
 import base.views.learning_units.update
 from attribution.views import attribution, tutor_application
 from base.views import learning_achievement, search, education_groups, user_list
@@ -138,17 +142,17 @@ urlpatterns = [
     ])),
 
     url(r'^learning_units/', include([
-        url(r'^by_activity/', base.views.learning_units.search.learning_units, name='learning_units'),
-        url(r'^by_service_course/', base.views.learning_units.search.learning_units_service_course,
+        url(r'^by_activity/', base.views.learning_units.search.simple.learning_units, name='learning_units'),
+        url(r'^by_service_course/', base.views.learning_units.search.service_course.learning_units_service_course,
             name='learning_units_service_course'),
-        url(r'^by_proposal/', base.views.learning_units.search.learning_units_proposal_search,
+        url(r'^by_proposal/', base.views.learning_units.search.proposal.learning_units_proposal_search,
             name='learning_units_proposal'),
-        url(r'^by_borrowed_course/', base.views.learning_units.search.learning_units_borrowed_course,
+        url(r'^by_borrowed_course/', base.views.learning_units.search.borrowed.learning_units_borrowed_course,
             name='learning_units_borrowed_course'),
         url(r'^by_summary/',
             base.views.learning_units.educational_information.LearningUnitDescriptionFicheSearch.as_view(),
             name='learning_units_summary'),
-        url(r'^by_external/', base.views.learning_units.search.learning_units_external_search,
+        url(r'^by_external/', base.views.learning_units.search.external.learning_units_external_search,
             name='learning_units_external'),
         url(r'^new/', include([
             url(r'^academic_year_id=(?P<academic_year_id>[0-9]+)$',
@@ -239,7 +243,7 @@ urlpatterns = [
         ])),
         url(r'^check/(?P<subtype>[A-Z]+)$', base.views.learning_units.common.check_acronym, name="check_acronym"),
     ])),
-    url(r'^proposals/search/$', base.views.learning_units.search.learning_units_proposal_search,
+    url(r'^proposals/search/$', base.views.learning_units.search.proposal.learning_units_proposal_search,
         name="learning_unit_proposal_search"),
 
     url(r'^my_osis/', include([
